# docs/swagger.yaml
definitions:
...
  models.OrganisationNew:
    properties:
      address:
        example: 123 Main St
        type: string
      description:
        example: My Organization Description
        type: string
      logo:
        example: https://www.example.com/logo.png
        type: string
      name:
        example: My Organization
        type: string
      phone:
        example: "1234567890"
        type: string
    required:
    - address
    - description
    - logo
    - name
    - phone
    type: object
...
  responses.OrganisationResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      message:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/organisation:
    post:
      consumes:
      - application/json
      description: Creates an organisation
      parameters:
      - description: Organisation data
        in: body
        name: organisation
        required: true
        schema:
          $ref: '#/definitions/models.OrganisationNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
      security:
      - BearerAuth: []
      summary: Create an organisation
      tags:
      - Organisation
  /api/organisation/{organisationId}:
    delete:
      consumes:
      - application/json
      description: Deletes an organisation
      parameters:
      - description: Organisation ID
        in: path
        name: organisationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
      security:
      - BearerAuth: []
      summary: Delete an organisation
      tags:
      - Organisation
    get:
      consumes:
      - application/json
      description: Gets an organisation
      parameters:
      - description: Organisation ID
        in: path
        name: organisationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
      security:
      - BearerAuth: []
      summary: Get an organisation
      tags:
      - Organisation
    put:
      consumes:
      - application/json
      description: Updates an organisation
      parameters:
      - description: Organisation ID
        in: path
        name: organisationId
        required: true
        type: string
      - description: Organisation object to be updated
        in: body
        name: organisation
        required: true
        schema:
          $ref: '#/definitions/models.OrganisationNew'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.OrganisationResponse'
      security:
      - BearerAuth: []
      summary: Update an organisation
      tags:
      - Organisation
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"